requirements.txt
asgiref==3.4.1
Django==3.1.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
